fix: resolve 30 broken imports on develop from module relocations
- nexus.rebac.* → nexus.bricks.rebac.* (factory/_system.py, nexus_fs.py, nexus_fs_core.py)
- nexus.lib.mutation_hooks → nexus.contracts.vfs_hooks (nexus_fs_bulk.py)
- nexus.core.vfs_hooks → nexus.contracts.vfs_hooks (vfs_hook_impls.py, nexus_fs_bulk.py, tests)
- nexus.core.metadata → nexus.contracts.metadata (test files)
- nexus.core.event_bus → nexus.core.file_events (nexus_fs_core.py, exporters, tests)
- nexus.storage.record_store_syncer → nexus.storage.record_store_write_observer (test_write_buffer.py)
- EnhancedReBACManager → ReBACManager (factory/_system.py, test_factory.py mock patches)
- Remove dead VFSHookPipeline TYPE_CHECKING import (class doesn't exist)
