+++ tests/e2e/self_contained/test_transactional_event_log.py
@@ -18,8 +18,8 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from nexus.core.event_bus import FileEventType
-from nexus.core.metadata import FileMetadata
+from nexus.contracts.metadata import FileMetadata
+from nexus.core.file_events import FileEventType
 from nexus.storage.models import OperationLogModel
 from nexus.storage.record_store import SQLAlchemyRecordStore
 from nexus.storage.record_store_write_observer import RecordStoreWriteObserver
