+++ tests/unit/storage/test_record_store_write_observer.py
@@ -16,7 +16,7 @@ import pytest
 from sqlalchemy.exc import OperationalError
 
 from nexus.contracts.exceptions import AuditLogError
-from nexus.core.metadata import FileMetadata
+from nexus.contracts.metadata import FileMetadata
 from nexus.storage.models import FilePathModel, OperationLogModel, VersionHistoryModel
 from nexus.storage.record_store import SQLAlchemyRecordStore
 from nexus.storage.record_store_write_observer import RecordStoreWriteObserver
