+++ tests/unit/storage/test_write_path_failures.py
@@ -15,7 +15,7 @@ from sqlalchemy import create_engine, event, select
 from sqlalchemy.orm import sessionmaker
 
 from nexus.contracts.exceptions import AuditLogError
-from nexus.core.metadata import DT_DIR, DT_REG, FileMetadata
+from nexus.contracts.metadata import DT_DIR, DT_REG, FileMetadata
 from nexus.storage.models import Base, FilePathModel
 from nexus.storage.record_store import RecordStoreABC
 from nexus.storage.record_store_write_observer import RecordStoreWriteObserver
